fix: patch stage_b inline HSIC sites + unify project_out to SVD + aggregate float32 site
Addresses reviewer-reported float32 overflow bug in HSIC denom computation.
Previous d82d0c1 fix missed the following sites:

stage_b_residualized_cka.py:
- Line 189-191: run_residualized_bootstrap null loop (3 HSIC sums)
- Line 238-240: run_residualized_a2_perm null loop (3 HSIC sums)
- Line 293: run_residualized_a2_perm_aggregate obs_hsic_xx init
- Line 306-307: run_residualized_a2_perm_aggregate null pair
- project_out_format_subspace: replaced inv(W@W.T) with SVD+rank
  threshold to handle rank-deficient 3-class multinomial LR weights
  (consistent with recompute_a2_bootstrap.py after af20cf7)

recompute_a2_bootstrap.py:
- Line 292: run_a2_perm_aggregate all_hsic_xx init missed float64 cast

tests/test_cka_overflow_e2e.py:
- New end-to-end overflow test: constructs n=252, d=6144 large-norm
  hidden states and runs full run_residualized_a2_perm_aggregate under
  np.seterr(over=raise, invalid=raise). Both medium (scale=30) and
  extreme (scale=100) cases pass with finite, nonzero CKA.

+++ scripts/recompute_a2_bootstrap.py
@@ -289,7 +289,7 @@ def main():
         for li in range(len(layers)):
             all_grams[li] = {fmt: all_fmt_data[li][fmt] @ all_fmt_data[li][fmt].T for fmt in FORMATS}
             all_centered[li] = {fmt: H_center @ all_grams[li][fmt] @ H_center for fmt in FORMATS}
-            all_hsic_xx[li] = {fmt: float(np.sum(all_centered[li][fmt] ** 2)) for fmt in FORMATS}
+            all_hsic_xx[li] = {fmt: float(np.sum(all_centered[li][fmt].astype(np.float64) ** 2)) for fmt in FORMATS}
             for f1, f2 in FORMAT_PAIRS:
                 obs_vals.append(_cka_from_centered(all_centered[li][f1], all_centered[li][f2]))
         obs_mean = float(np.mean(obs_vals))
